Create .__pre-orig.tmux_variables.sh.bak
Creating a reference point before the start-of-git clean-up, getting rid of extra comments, etc. 

This is my slightly-modified version of
`<root-of-tmux-stuff>/plugins/tmux-logging/scripts/variables.sh', 
which allows for the auto-logging of TMUX terminals in just the way I like.
The path of the `variables.sh' script can also be represented
`<root-of-tmux-logging>/scripts/variables.sh'
and is the following path on my machine
`~/.tmux/plugins/tmux-logging/scripts/variables.sh'

Note that this file needs to be modified to get the tailored filename that I want. If a new version of TMUX comes out and I update the files, I'll need to make the same changes to the new `variables.sh'