Create .__pre-orig.vimrc
Part of creating a reference point of things as they were before I started putting all these reference files and dotfiles in git. This file is the .vimrc that comes when you do `vimtutor', but it has some changes I've already put in, specifically for indentation/spacing/tabs. This will get changed as I put more vim plugins here. I have a bunch of youtube videos and iggredible's tutorials. I should find the links, and I hope to do so.